[feat] backend basic crud setup for provider

+++ backend-web-go/models/providers.go
@@ -6,7 +6,7 @@ type Provider struct {
 	ID        uint      `json:"id" gorm:"primary_key"`
 	Title     string    `json:"title" gorm:"type:text"`
 	Image     string    `json:"image" gorm:"type:varchar(100)"`
-	Order     int       `json:"order" gorm:"type:integer"`
+	Order     int16     `json:"order" gorm:"type:integer"`
 	CreatedAt time.Time `json:"created_at" gorm:"default.CURRENT_TIMESTAMP"`
 	UpdatedAt time.Time `json:"updated_at" gorm:"default.CURRENT_TIMESTAMP"`
 }
